fix: changed size_t to int to avoid narrowing converstion errors

+++ ANN/CLData.h
@@ -54,7 +54,7 @@ const char* cl_errstr(cl_int error);
 void cl_print_err(const char* entry, cl_int error);
 // Allocate a cl buffer with data in it, NULL for 0 init
 cl_mem alloc_buffer(cl_context context, const char* name, 
-    size_t size, void* data=NULL, cl_mem_flags flag=CL_MEM_READ_WRITE);
+    int size, void* data=NULL, cl_mem_flags flag=CL_MEM_READ_WRITE);
 
 Function derivative(Function f);
 const char* function_to_string(Function f);
@@ -67,9 +67,9 @@ const char* function_arg_string(Function f);
 // All cl_mem should be given as pointers.
 void call_kernel(CLdata* cl, Function fun, 
                  cl_uint         work_dim,
-                 const size_t*   global_work_offset,
-                 const size_t*   global_work_size,
-                 const size_t*   local_work_size,
+                 const int*      global_work_offset,
+                 const int*      global_work_size,
+                 const int*      local_work_size,
                  cl_uint         num_events_in_wait_list,
                  const cl_event* event_wait_list,
                  cl_event*       event ...);
